Addressed all comments in PR
- Throws error and exists if file name does not exist in 'singleupload.js' and 'bulkupload.js'
- overwrites collection only if it exists, otherwise just creates new collection
- inserted comments
- changed input reader's variable name from 'rl' to 'input_reader'

+++ node.js_mongodb/SingleUpload.js
@@ -2,23 +2,35 @@ const DBHelper=require('./DB_helper');
 const readline = require('readline');
 PromptAndUpload();
 
+/* Adds a new collection if it does not exist. If collection exists, replaces it with the given collection */
 async function PromptAndUpload() {
-    const rl = readline.createInterface({
+    const input_reader = readline.createInterface({
         input: process.stdin,
         output: process.stdout
       });
       
-      rl.question('File name: ', async (jsonFileName) => {
-        rl.question('Collection name: ', async (collectionName) => {
-            console.log("Overwriting data..");
+      input_reader.question('File name: ', async (jsonFileName) => {
+        var all_data;
+        try{
+            all_data = require('./data/'+jsonFileName);
+        }catch(e){
+            console.error('File not found');
+            return null;
+        }
+        input_reader.question('Collection name: ', async (collectionName) => {
             let db="rsquared_zilla_upazilla_db";
-            var all_data = require('./data/'+jsonFileName);
-            let deleteExisting= await DBHelper.deleteData(db,collectionName,{});
-            console.log("Existing data from "+collectionName+" deleted.", deleteExisting);
-            var newData_id=await DBHelper.addData(db,collectionName,all_data);
+            var allExistingCollections=await DBHelper.getAllCollectionNames();
+            if(allExistingCollections.indexOf(collectionName)>=0){
+                console.log("Overwriting data..");
+                let deleteExisting= await DBHelper.deleteData(db,collectionName,{});
+                console.log("Existing data from "+collectionName+" deleted.", deleteExisting);
+            }else{
+                console.log("Creating new collection: ", collectionName);
+            }
+            await DBHelper.addData(db,collectionName,all_data);
             console.log("Data added to collection '"+collectionName+"' from '"+jsonFileName+"'");
             console.log("TASK COMPLETE");
-            rl.close();
+            input_reader.close();
         });
       });
 }
